# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlobeVote Civic Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /countries:
    get:
      operationId: listCountries
      tags: [countries]
      summary: List all countries with election info
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
  /countries/{id}:
    get:
      operationId: getCountry
      tags: [countries]
      summary: Get country details with voting guide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Country detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryDetail"

  # ELECTIONS
  /elections:
    get:
      operationId: listElections
      tags: [elections]
      summary: List upcoming and recent elections
      parameters:
        - name: countryId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, completed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of elections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Election"
  /elections/{id}:
    get:
      operationId: getElection
      tags: [elections]
      summary: Get detailed election info and results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Election detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionDetail"
...
  /electoral-systems/{id}:
    get:
      operationId: getElectoralSystem
      tags: [electoral-systems]
      summary: Get detailed electoral system info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Electoral system detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectoralSystemDetail"
  /electoral-systems/simulate:
    post:
      operationId: simulateElection
      tags: [electoral-systems]
      summary: Simulate seat allocation with given vote shares
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationRequest"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"

  # AI CHAT
  /chat/ask:
    post:
      operationId: askCivicAssistant
      tags: [chat]
      summary: Ask the AI civic assistant a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get recent chat history
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  # QUIZ / GAMIFICATION
...
  /claims/verify:
    post:
      operationId: verifyClaim
      tags: [claims]
      summary: Verify a civic claim or piece of election information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimVerifyRequest"
      responses:
        "200":
          description: Claim verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimVerifyResult"
  /claims/recent:
    get:
      operationId: getRecentClaims
      tags: [claims]
      summary: Get recently verified claims
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClaimVerifyResult"

  # NEWS
...
components:
  schemas:
...
    Country:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        flag:
          type: string
        region:
          type: string
        electoralSystem:
          type: string
        nextElectionDate:
          type: string
          nullable: true
        voterTurnout:
          type: number
          nullable: true
        registeredVoters:
          type: integer
          nullable: true
      required: [id, name, code, flag, region, electoralSystem]
    CountryDetail:
      allOf:
        - $ref: "#/components/schemas/Country"
        - type: object
          properties:
            votingAge:
              type: integer
            registrationDeadlineDays:
              type: integer
            idRequirements:
              type: array
              items:
                type: string
            votingProcess:
              type: array
              items:
                type: string
            overview:
              type: string
            keyFacts:
              type: array
              items:
                type: string
          required: [votingAge, registrationDeadlineDays, idRequirements, votingProcess, overview, keyFacts]
    Election:
      type: object
      properties:
        id:
          type: string
        countryId:
          type: string
        countryName:
          type: string
        countryFlag:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [presidential, parliamentary, local, referendum, midterm]
        date:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        description:
          type: string
      required: [id, countryId, countryName, countryFlag, title, type, date, status, description]
    ElectionDetail:
      allOf:
        - $ref: "#/components/schemas/Election"
        - type: object
          properties:
            results:
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/ElectionResult"
            candidates:
              type: array
              items:
                $ref: "#/components/schemas/Candidate"
            turnout:
              type: number
              nullable: true
            totalSeats:
              type: integer
              nullable: true
          required: [candidates]
    ElectionResult:
      type: object
      properties:
        partyName:
          type: string
        partyColor:
          type: string
        voteShare:
          type: number
        seats:
          type: integer
          nullable: true
      required: [partyName, partyColor, voteShare]
    Candidate:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        partyColor:
          type: string
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
      required: [name, party, partyColor, description]
...
    ElectoralSystem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        abbreviation:
          type: string
        description:
          type: string
        countriesUsing:
          type: integer
        fairnessScore:
          type: integer
        simplicityScore:
          type: integer
        proportionalityScore:
          type: integer
        examples:
          type: array
          items:
            type: string
      required: [id, name, abbreviation, description, countriesUsing, fairnessScore, simplicityScore, proportionalityScore, examples]
    ElectoralSystemDetail:
      allOf:
        - $ref: "#/components/schemas/ElectoralSystem"
        - type: object
          properties:
            howItWorks:
              type: array
              items:
                type: string
            pros:
              type: array
              items:
                type: string
            cons:
              type: array
              items:
                type: string
            historyAndOrigin:
              type: string
          required: [howItWorks, pros, cons, historyAndOrigin]
    SimulationRequest:
      type: object
      properties:
        systemId:
          type: string
        totalSeats:
          type: integer
        parties:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              voteShare:
                type: number
              color:
                type: string
            required: [name, voteShare, color]
        threshold:
          type: number
          default: 5
      required: [systemId, totalSeats, parties]
    SimulationResult:
      type: object
      properties:
        systemId:
          type: string
        totalSeats:
          type: integer
        parties:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              voteShare:
                type: number
              seats:
                type: integer
              seatShare:
                type: number
              color:
                type: string
            required: [name, voteShare, seats, seatShare, color]
        gallagherIndex:
          type: number
        summary:
          type: string
      required: [systemId, totalSeats, parties, gallagherIndex, summary]
    ChatRequest:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: string
        mode:
          type: string
          enum: [standard, eli5, expert]
          default: standard
        countryContext:
          type: string
          nullable: true
      required: [message, sessionId]
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        sources:
          type: array
          items:
            type: string
        confidence:
          type: number
        followUpSuggestions:
          type: array
          items:
            type: string
        sessionId:
          type: string
        createdAt:
          type: string
      required: [id, message, sources, confidence, followUpSuggestions, sessionId, createdAt]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        confidence:
          type: number
          nullable: true
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
...
    ClaimVerifyRequest:
      type: object
      properties:
        claim:
          type: string
        context:
          type: string
          nullable: true
        countryId:
          type: string
          nullable: true
      required: [claim]
    ClaimVerifyResult:
      type: object
      properties:
        id:
          type: string
        claim:
          type: string
        verdict:
          type: string
          enum: [true, mostly-true, mixed, mostly-false, false, unverifiable]
        confidence:
          type: number
        explanation:
          type: string
        sources:
          type: array
          items:
            type: string
        manipulationFlags:
          type: array
          items:
            type: string
        verifiedAt:
          type: string
      required: [id, claim, verdict, confidence, explanation, sources, manipulationFlags, verifiedAt]
...