# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlobeVote Civic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestMode } from "./chatRequestMode";

export interface ChatRequest {
  message: string;
  sessionId: string;
  mode?: ChatRequestMode;
  countryContext?: string | null;
}