# lib/api-zod/src/generated/types/claimVerifyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlobeVote Civic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimVerifyRequest {
  claim: string;
  context?: string | null;
  countryId?: string | null;
}