# lib/api-zod/src/generated/types/electionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlobeVote Civic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ElectionResult {
  partyName: string;
  partyColor: string;
  voteShare: number;
  seats?: number | null;
}